+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to semantic versioning for configuration changes.
 
 ## [Unreleased]
 
+### Auto-updated - 2026-01-10 07:15:06
+- Modified: scripts/test_net2_signalr.py
+
+## [Unreleased]
+
 ### Auto-updated - 2026-01-10 07:00:15
 - Added: scripts/test_net2_signalr.py